Added mechanism to "reattach" to processes that had been started earlier.
In practice this doesn’t seem useful as-is because the children die if SIGINT is sent to parent.  However, could change children to ignore signal and allow re-attaching.

However, some improvements to logging/error messages were also made.  Eventually could support flock as a backend which will have very long running processes which we will want to be able to re-attach if something goes wrong.