Enable FK constraint check as soon as we connect to the DB
This is being enabled to discover FK mistakes which are currently silently arising.

Specifically, at the moment we've seen a problem where something went very wrong when conseq was running and the host was unexpectantly shut down.

Records in rule_execution were marked as "complete" but the corresponding records in execution where missing. Could be a transaction boundary issue or
an issue with the GC command. It's unclear at this point, but at least enforcing FKs will fail the operation which attempts to put the DB into an invalid state.