Changes around dataflow, rule filters, new docs
This commit ended up being a fair number of changes which were made in the course of using conseq and wanting to
streamline the common tasks.

WIP: dataflow now contains a new execution loop which is based on a static analysis of the rules. However, this is not yet finished.
parser now captures filename and line numbers for each rule definition which could be used in reporting.
Some formatting improvements to the the html report