A fairly extensive rewrite of "conseq stage" command
Basically changed the approach. Instead of simulating
a run and keeping track of which rules would
have consumed which artifacts, just run the
queries manually.

Also made fixes for problems encountered when
testing on a complicated example.

SCRIPT_DIR now is always relative to the
file that defined the rule the variable is using.